backend/pyproject.toml
[project]
name = "medicallm-backend"
version = "0.1.0"
description = "MedicaLLM - Medical AI Assistant Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.0",
    "einops>=0.8.1",
    "python-dotenv>=1.0.0",
    "fastapi>=0.119.1",
    "langchain>=1.0.2",
    "langchain-aws>=0.2.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-huggingface>=1.0.1",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf>=6.3.0",
    "sentence-transformers>=5.1.2",
    "icecream>=2.1.8",
    "uvicorn[standard]>=0.38.0",
    "xsdata[cli,lxml]>=25.7",
    "pymed>=0.8.9",
    "tqdm>=4.66.0",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
    "email-validator>=2.0.0",
    "cachetools>=5.3.0",
    "slowapi>=0.1.9",
]